change raster name to end with '_intensity' instead of no suffix for intensity grid; increase IntensitySize to uint32_t to prevent overflow

+++ tbd/src/cpp/stdafx.h
@@ -170,7 +170,7 @@ using SlopeSize = uint16_t;
 /**
  * \brief Type used for storing intensities
  */
-using IntensitySize = uint16_t;
+using IntensitySize = uint32_t;
 /**
  * \brief Type used for storing distances within cells
  */
